+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Add `Script::values()` to iterate over all scripts.
 * Add `Lang::values()` to iterate over all languages.
 * Add integration with `enum-map`
+* Implement `FromStr` for `Script` and `Lang`
 
 ### v0.11.1 - 2020-11-28
 * Use Trigram tuple instead of heap allocated String. (~68% faster)